Define separate class for memory and object management, safe_repr
Previously printing the GloomObject.objects dict incremented the refcount
for each object in the dictionary individually, and calculated the global objects
total for each item. This meant the statistics never matched if there were
multiple items in the dict.

This change adds a safe_repr which doesn't impact the GloomObject state on invocation
We use that safe_repr when building the GloomObject.object __repr__ string