Make things somewhat less convoluted
Previously we had GloomObjects and also GloomValues (?) and
GloomTableObjects and GloomNumberObjects and GloomStringObjects and
TableValues (???). Now we just have GloomObjects. Thank god.

In Gloom, "everything is an object", so I would like for the properties
of objects to be represented as table-valued Gloom objects, but I think
the current abstractions won't allow that, so they aren't quite right
yet.

In Python, every e.g. PyLongObject or PyDictObject can be cast to a
PyObject*, which is the first member of the Py<T>Object struct. This
allows for dynamic ops on the C representations of Python objects. One
thing to keep in mind is that PyObjects are not completely analagous to our
Gloom objects, since PyObjects don't actually carry any object data --
except for type information about the underlying object, along with refcounts

However, since it's the first member of the underlying Py<T>Object, it acts
as a sort of header for the object, allowing e.g. functions to determine if
casts are needed to make operations work

All that is to say that we will probably end up decomposing things again!!